beta version ended

+++ src/Controller/AdminController.php
@@ -5,8 +5,10 @@ namespace App\Controller;
 use App\Entity\Student;
 use App\Form\StudentType;
 use App\Repository\StudentRepository;
+use Doctrine\DBAL\DBALException;
 use Doctrine\ORM\EntityManagerInterface;
 use FOS\RestBundle\Request\ParamFetcher;
+use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\Request;
@@ -25,13 +27,14 @@ class AdminController extends AbstractController
     {
         $this->entityManager = $entityManager;
     }
+
     /**
      * @Route("/", name="student_index", methods={"GET"})
      */
-    public function index(StudentRepository $studentRepository): Response
+    public function index(Request $request, StudentRepository $studentRepository): Response
     {
         return $this->render('student/index.html.twig', [
-            'students' => $studentRepository->findAll(),
+            'students' => $studentRepository->findBy(['user_id' => $this->getUser()->getId()])
         ]);
     }
 
@@ -39,13 +42,14 @@ class AdminController extends AbstractController
      * @Route("/uploadStudents", name="student_upload", methods={"POST"})
      * @Rest\FileParam(name="uploadedfile", nullable=false)
      */
-    public function uploadStudents(ParamFetcher $paramFetcher)
+    public function uploadStudents(ParamFetcher $paramFetcher, LoggerInterface $logger)
     {
         /** @var UploadedFile $file */
         $file = $paramFetcher->get('uploadedfile');
         $file_content = json_decode(file_get_contents($file->getPathname()));
+        $user_id = $this->getUser()->getId();
 
-        foreach($file_content as $student_raw) {
+        foreach ($file_content as $student_raw) {
             $student = new Student();
             $student->setAlbumNo($student_raw->album_no);
             $student->setName($student_raw->name);
@@ -53,11 +57,19 @@ class AdminController extends AbstractController
             $student->setStartYear($student_raw->start_year);
             $student->setSemester($student_raw->semester);
             $student->setCardUid($student_raw->card_uid);
+            $student->setUserId($user_id);
 
             $this->entityManager->persist($student);
         }
 
-        $this->entityManager->flush();
+        try {
+            $this->entityManager->flush();
+        } catch (DBALException $e) {
+
+            $this->addFlash('danger', 'Error occured during storing students in DB');
+            $logger->error($e->getMessage());
+            return $this->redirectToRoute('student_index');
+        }
         return $this->redirectToRoute('student_index');
     }
 
@@ -67,13 +79,16 @@ class AdminController extends AbstractController
     public function new(Request $request): Response
     {
         $student = new Student();
-        $form = $this->createForm(StudentType::class, $student);
+        $form = $this->createForm(  StudentType::class, $student);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($student);
+            $student->setUserId($this->getUser()->getId());
             $entityManager->flush();
+
+            $this->addFlash('success', 'Student created successfully');
 
             return $this->redirectToRoute('student_index');
         }
@@ -89,6 +104,9 @@ class AdminController extends AbstractController
      */
     public function show(Student $student): Response
     {
+        if(!$this->checkOwnership($student)) {
+            return $this->redirect('main');
+        }
         return $this->render('student/show.html.twig', [
             'student' => $student,
         ]);
@@ -99,6 +117,9 @@ class AdminController extends AbstractController
      */
     public function edit(Request $request, Student $student): Response
     {
+        if(!$this->checkOwnership($student)) {
+            return $this->redirect('main');
+        }
         $form = $this->createForm(StudentType::class, $student);
         $form->handleRequest($request);
 
@@ -115,10 +136,13 @@ class AdminController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="student_delete", methods={"DELETE"})
+     * @Route("/{id}/delete", name="student_delete", methods={"DELETE"})
      */
     public function delete(Request $request, Student $student): Response
     {
+        if(!$this->checkOwnership($student)) {
+            return $this->redirect('main');
+        }
         if ($this->isCsrfTokenValid('delete' . $student->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($student);
@@ -126,5 +150,15 @@ class AdminController extends AbstractController
         }
 
         return $this->redirectToRoute('student_index');
+    }
+
+    private function checkOwnership(Student $student) : bool
+    {
+        if($this->getUser()->getId() !== $student->getUserId()) {
+            $this->addFlash('danger', 'You have to be owner of student');
+            return false;
+        }
+
+        return true;
     }
 }
